Remove unused NodeFront/ValueFront methods
